[Test] 상세레시피 조회로직 테스트코드 구현

+++ src/main/java/com/example/foodking/Emotion/EmotionService.java
@@ -36,7 +36,9 @@ public class EmotionService {
         }
         else{
             ReplyEmotion replyEmotion = result.get();
-            isMyEmotion(user.getUserId(),replyEmotion.getUser());
+            if(!isMyEmotion(user.getUserId(),replyEmotion.getUser()))
+                throw new CommondException(ExceptionCode.ACCESS_FAIL_EMOTION);
+
             replyEmotionRepository.delete(result.get());
         }
     }
@@ -54,7 +56,9 @@ public class EmotionService {
         }
         else{
             RecipeEmotion recipeEmotion = result.get();
-            isMyEmotion(user.getUserId(),recipeEmotion.getUser());
+            if(!isMyEmotion(user.getUserId(),recipeEmotion.getUser()))
+                throw new CommondException(ExceptionCode.ACCESS_FAIL_EMOTION);
+
             recipeEmotionRepository.delete(recipeEmotion);
         }
     }
@@ -67,14 +71,10 @@ public class EmotionService {
         return recipeEmotionRepository.countByRecipeInfoAndEmotionType(recipeInfo,EmotionType.Like);
     }
 
-    public void isMyEmotion(Long userId, User user){
-        /*
-           단위 테스트시 userId값을 지정할 수 없기때문에 해당 조건문을 추가하여 테스트를 통과할 수 있도록 했다.
-           실제 환경에서는 User는 null이 아니고 user.userId값은 null인 경우는 존재하지 않는다.
-        */
-        if(user != null && user.getUserId() == null)
-            ;
-        else if( user ==null || !userId.equals(user.getUserId()) )
-            throw new CommondException(ExceptionCode.ACCESS_FAIL_EMOTION);
+    public boolean isMyEmotion(Long userId, User user){
+        if( user ==null || !userId.equals(user.getUserId()) )
+            return false;
+
+        return true;
     }
 }
